rename event filter components
