+++ apps/web/src/app/arrangementer/EventListPage.tsx
@@ -36,11 +36,11 @@ import { CalendarWeekNavigation } from "./components/calendar/EventWeekCalendar/
 import { EventWeekCalendar } from "./components/calendar/EventWeekCalendar/EventWeekCalendar"
 import { EventList, EventListSkeleton } from "./components/EventList"
 import { RegisteredEventsCard } from "./components/RegisteredEventsCard"
-import { EventListFilterChips } from "./components/filters/EventListFilterChips"
-import { GroupFilter } from "./components/filters/GroupFilter"
+import { EventListFilterChips } from "./components/filters/EventFilterChips"
+import { EventGroupFilter } from "./components/filters/EventGroupFilter"
 import { SearchInput } from "@/components/molecules/ListFilters/SearchInput"
-import { SortFilter } from "./components/filters/SortFilter"
-import { TypeFilter } from "./components/filters/TypeFilter"
+import { EventSortFilter } from "./components/filters/EventSortFilter"
+import { EventTypeFilter } from "./components/filters/EventTypeFilter"
 import {
   useEventAllSummariesByAttendingUserIdInfiniteQuery,
   useEventAllSummariesInfiniteQuery,
@@ -313,13 +313,13 @@ export const EventListPage = ({ initialListViewMode }: Props) => {
                           <Text element="span" className="h-5.5 font-medium text-sm">
                             Sorter
                           </Text>
-                          <SortFilter
+                          <EventSortFilter
                             value={filters.viewModeSort}
                             onChange={(viewModeSort) => updateFilters({ viewModeSort })}
                           />
                         </div>
                         <div className="mt-6">
-                          <TypeFilter
+                          <EventTypeFilter
                             value={filters.types}
                             onChange={(types) => updateFilters({ types })}
                             isStaff={isStaff}
@@ -328,7 +328,7 @@ export const EventListPage = ({ initialListViewMode }: Props) => {
                       </div>
 
                       <div className="mt-6 sm:mt-0">
-                        <GroupFilter
+                        <EventGroupFilter
                           value={filters.groups}
                           onChange={(groups) => updateFilters({ groups })}
                           groups={groups ?? []}
@@ -354,7 +354,7 @@ export const EventListPage = ({ initialListViewMode }: Props) => {
                 className="max-sm:hidden w-full max-w-90"
               />
 
-              <SortFilter
+              <EventSortFilter
                 value={filters.viewModeSort}
                 onChange={(viewModeSort) => updateFilters({ viewModeSort })}
                 className="max-md:hidden"
@@ -409,16 +409,19 @@ export const EventListPage = ({ initialListViewMode }: Props) => {
 
       {isEventListView && searchBarOpen && (
         <div className="sm:hidden">
-          <SearchInput initialValue={filters.search} onDebouncedChange={(value) => updateFilters({ search: value })} />
+          <EventSearchInput
+            initialValue={filters.search}
+            onDebouncedChange={(value) => updateFilters({ search: value })}
+          />
         </div>
       )}
 
       {isEventListView && (
         <div className="md:grid md:grid-cols-[15rem_auto] md:gap-8 lg:gap-12 min-w-0">
           <div className="max-md:hidden mt-4">
-            <TypeFilter value={filters.types} onChange={(types) => updateFilters({ types })} isStaff={isStaff} />
+            <EventTypeFilter value={filters.types} onChange={(types) => updateFilters({ types })} isStaff={isStaff} />
             <div className="mt-6">
-              <GroupFilter
+              <EventGroupFilter
                 value={filters.groups}
                 onChange={(groups) => updateFilters({ groups })}
                 groups={groups ?? []}
